requirements.txt
openai==1.101.0
pymupdf==1.26.4
PyPDF2==3.0.1
python-dotenv==1.1.0
tiktoken==0.11.0
pyyaml==6.0.2
openrouter==0.1.3
streamlit==1.39.0
pytest==8.3.4
pytest-asyncio==0.24.0
